Style fixes
Add labelh_new_user() function that allocates a label handle and a
normal ipc port in the specified task's address space.  The specified
label is inserted into the label handle and a send/receive right
for the label handle's port is inserted into the task's space.  The
port name is passed back in the namep parameter.  Unlike labelh_new()
the specified task now holds the receive right for the port which
means that when the task dies (or calls mach_port_destroy()), the
port and label handle will be garbage collected.

Use labelh_new_user() in mac_label_new().

Use labelh_new() in labelh_duplicate().